cleanup stitching package tests.
