+++ packages/stitch/tests/transforms.test.ts
@@ -12,6 +12,12 @@ import { stitchSchemas } from '../src/stitchSchemas';
 
 import { propertySchema } from './fixtures/schemas';
 
+import {
+  RemoveFieldDirectives,
+  RemoveFieldsWithDirective
+} from '../src/index';
+
+
 describe('rename root type', () => {
   test('works with stitchSchemas', async () => {
     let schemaWithCustomRootTypeNames = makeExecutableSchema({
@@ -126,5 +132,72 @@ describe('filter fields', () => {
     });
 
     assertValidSchema(stitchedSchema);
+  });
+});
+
+describe('RemoveFieldsWithDirective', () => {
+  test('works', async () => {
+    const listingsSchema = makeExecutableSchema({
+      typeDefs: `
+        type Listing {
+          id: ID!
+          description: String!
+          price: Float! @deprecated(reason: "other deprecation")
+          sellerId: ID! @deprecated(reason: "stitching use only")
+          buyerId: ID  @deprecated(reason: "stitching use only")
+        }
+      `
+    });
+
+    const stitchedSchema = stitchSchemas({
+      subschemas: [
+        {
+          schema: listingsSchema,
+          transforms: [new RemoveFieldsWithDirective('deprecated', { reason: 'stitching use only' })]
+        }
+      ],
+    });
+
+    expect(listingsSchema.getType('Listing').getFields().price.deprecationReason).toBe('other deprecation');
+    expect(listingsSchema.getType('Listing').getFields().sellerId.deprecationReason).toBe('stitching use only');
+
+    expect(stitchedSchema.getType('Listing').getFields().price.deprecationReason).toBe('other deprecation');
+    expect(stitchedSchema.getType('Listing').getFields().sellerId).toBeUndefined();
+  });
+});
+
+describe('RemoveFieldDirectives', () => {
+  test('works', async () => {
+    const usersSchema = makeExecutableSchema({
+      typeDefs: `
+        type User {
+          id: ID!
+          email: String! @deprecated(reason: "other deprecation")
+        }
+        type Listing {
+          seller: User! @deprecated(reason: "gateway access only")
+          buyer: User @deprecated(reason: "gateway access only")
+        }
+      `
+    });
+
+    const stitchedSchema = stitchSchemas({
+      subschemas: [
+        {
+          schema: usersSchema,
+          transforms: [new RemoveFieldDirectives('deprecated', { reason: 'gateway access only' })]
+        },
+      ],
+    });
+
+    expect(usersSchema.getType('Listing').getFields().seller.deprecationReason).toBe('gateway access only');
+    expect(usersSchema.getType('Listing').getFields().seller.astNode.directives.length).toEqual(1);
+    expect(stitchedSchema.getType('Listing').getFields().seller.deprecationReason).toBeUndefined();
+    expect(stitchedSchema.getType('Listing').getFields().seller.astNode.directives.length).toEqual(0);
+
+    expect(usersSchema.getType('User').getFields().email.deprecationReason).toBe('other deprecation');
+    expect(usersSchema.getType('User').getFields().email.astNode.directives.length).toEqual(1);
+    expect(stitchedSchema.getType('User').getFields().email.deprecationReason).toBe('other deprecation');
+    expect(stitchedSchema.getType('User').getFields().email.astNode.directives.length).toEqual(1);
   });
 });
